Moved hello.py to app.py. Added gunicorn run config to procfile.
